Make Python 3 based installs work
This moves to my fork of puppet-python as it contains a fix for
working with virtualenv >= 20. Otherwise this is just poking
puppet-python to cope with Fedora 33 *only* having Python 3,
which it makes it available as both `python` and `python3`.

See https://github.com/voxpupuli/puppet-python/pull/578 for the
aforementioned virtualenv fix.